requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
google-generativeai>=0.8.3
python-dotenv==1.0.1
pydantic==2.10.4
python-multipart==0.0.18
httpx==0.28.1
celery==5.4.0
redis==5.2.1
